moved prepareZentra function into zentra.py.
